lots of bugfixes - changed to data loader - added autofocus - fixed deletion of whiteboards - fixed download of whiteboards - fixed upload of whiteboards

+++ app/src/main/java/co/whiteboardmaster/android/model/ThumbImageMessage.java
@@ -12,7 +12,7 @@ public class ThumbImageMessage {
 
     public static int THUMB_IMAGE = 1;
 
-    public int imageId;
+    public long imageId;
     public int destHeight;
     public int destWidth;
     public ImageView imageView;
